Renamed `format` to `articleFormat`
