+++ apps-rendering/src/articleFormat.ts
