+++ apps-rendering/src/components/Card/index.tsx
@@ -40,7 +40,7 @@ import Img from 'components/ImgAlt';
 import Kicker from 'components/Kicker';
 import { stars } from 'components/StarRating';
 import { formatSeconds, makeRelativeDate } from 'date';
-import { getPillarOrElseNews } from 'format';
+import { getPillarOrElseNews } from 'articleFormat';
 import type { Image } from 'image';
 import { maybeRender, pipe } from 'lib';
 import type { FC, ReactElement } from 'react';
