+++ apps-rendering/src/renderer.ts
@@ -66,7 +66,7 @@ import OrderedList from 'components/OrderedList';
 import Paragraph from 'components/Paragraph';
 import Pullquote from 'components/Pullquote';
 import RichLink from 'components/RichLink';
-import { themeToPillar } from 'format';
+import { themeToPillar } from 'articleFormat';
 import { isElement, pipe } from 'lib';
 import { createElement as h } from 'react';
 import type { ReactElement, ReactNode } from 'react';
